Added damage and speed attributes to enhance via the experience/leveling screen. The experience screen now draws bars to indicate how many levels in an attribute you have invested. They max out at 5. Modified other classes to compensate for new attributes. Current skill points are drawn at the bottom of the experience/leveling screen.

+++ GenericZombieShooter/src/genericzombieshooter/actors/Zombie.java
@@ -36,7 +36,7 @@ public class Zombie extends Point2D.Double {
     private Animation img;
     
     private int type;
-    private int health; // How much health the zombie has.
+    private double health; // How much health the zombie has.
     private int damage; // How much damage the zombie does per tick to the player.
     private double speed; // How fast the zombie moves.
     private int cashValue; // How many points the zombie is worth.
@@ -60,7 +60,7 @@ public class Zombie extends Point2D.Double {
         // To be overridden.
     }
     public int getType() { return this.type; }
-    public int getHealth() { return this.health; }
+    public double getHealth() { return this.health; }
     public void takeDamage(int damage_) { this.health -= damage_; }
     public int getDamage() { return this.damage; }
     public int getCashValue() { return this.cashValue; }
